requirements.txt
aiofiles==22.1.0
aiosqlite==0.17.0
anyio==3.6.2
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
da-vinci==0.3.0
dashboard==0.0.6
databases==0.6.2
Django==4.1.2
django-admin-thumbnails==0.2.6
django-thumbnails==0.7.0
greenlet==2.0.1
h11==0.14.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.1
orm==0.2.0.dev1
Pillow==9.3.0
psycopg2==2.9.5
python-multipart==0.0.5
requests==2.28.1
shortuuid==1.0.11
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlparse==0.4.3
starlette==0.21.0
typesystem==0.3.0.dev0
typing_extensions==4.4.0
tzdata==2022.6
urllib3==1.26.12
uvicorn==0.19.0
